+++ tools/emitRedraw/tool.js
@@ -42,6 +42,7 @@
     async start() {
       const lt = this.lichessTools;
       const lichess = lt.lichess;
+      if (!lichess || !lt.uiApi) return;
       const $ = lt.$;
       let emit = null;
       emit = lt.debounce(() => {